Added usage and instructions to build the package.
Added more up to date description, instructions about how to build the package and dependencies.

lxc-duply-backup uses an early version of basalt (https://github.com/basalt/python-basalt) to build the packages. Since basalt is not ready to ship yet, the README has been expanded with the full set of dependencies. In a later version this can be replaced with basalt.